+++ packages/python/openreceive/src/openreceive/_generated/tables.py
@@ -115,6 +115,7 @@ SWAP_ASSET_INFO: Final = {
         "network_label": "Tron",
         "coin": "USDT",
         "network": "TRX",
+        "pegged_to": "USD",
     },
     "USDT_SOL": {
         "pay_in_asset": "USDT_SOL",
@@ -122,6 +123,7 @@ SWAP_ASSET_INFO: Final = {
         "network_label": "Solana",
         "coin": "USDT",
         "network": "SOL",
+        "pegged_to": "USD",
     },
     "USDC_SOL": {
         "pay_in_asset": "USDC_SOL",
@@ -129,6 +131,7 @@ SWAP_ASSET_INFO: Final = {
         "network_label": "Solana",
         "coin": "USDC",
         "network": "SOL",
+        "pegged_to": "USD",
     },
     "ETH_ETH": {
         "pay_in_asset": "ETH_ETH",
@@ -143,6 +146,7 @@ SWAP_ASSET_INFO: Final = {
         "network_label": "Ethereum",
         "coin": "USDT",
         "network": "ETH",
+        "pegged_to": "USD",
     },
     "USDC_ETH": {
         "pay_in_asset": "USDC_ETH",
@@ -150,6 +154,7 @@ SWAP_ASSET_INFO: Final = {
         "network_label": "Ethereum",
         "coin": "USDC",
         "network": "ETH",
+        "pegged_to": "USD",
     },
 }